tests: make the certificate fixtures safe to share, and give them their own file
TestCert lived inside TestServer.cs and had three problems that all produce
failures pointing somewhere other than the fixture.

Five suites share /tmp/ioxide-e2e-mtls and /tmp/ioxide-e2e-tls with nothing
serialising creation. Two processes minting at once interleave their writes to
ca.crt and ca.key, so a certificate from one generation ends up beside a key from
another - every handshake using the pair fails, and it reads as a bug in whatever
was under test. Creation now holds a cross-process lock for the whole
check-and-mint, so the check cannot succeed while another process is still
writing the files being checked.

Writes went straight to the final path, so a reader could see a half-written PEM.
Each file is now written to a temporary name and renamed into place.

And the cache never expired while leaves are minted for a year, so a checkout
older than that failed every test expecting a handshake to succeed and passed
every test expecting one to fail - the worse direction. The freshness check now
loads the certificate and checks its validity window. It also checks every file
rather than a sample: the mutual-TLS fixture verified four of its eight paths, so
a partially evicted cache handed back paths to files that were not there.

Verified by deleting both fixture directories and running all five suites
concurrently from cold - the case that used to produce mismatched pairs. 287
tests green across the five.
